Major update to `parse_blocks` inside of `_.parse`
This update was to seperate out what's happening in this file into individual functions to make it clearer as to what's going on. Also seperated out `this.merge` into two seperate functions since the first half of `this.merge` was actually focused on calling the annotation function.

This will allow me to implement the `///#{2} @remove` for lines inside of the code block.