Overhauled `get_blocks` to be a reusable function
Before it was multiple for loops, to get both header and body blocks. This also fixes a few edge cases, where there might only be comments in a file and that's it, or if a file is empty, etc.(see tests/lib/edge-cases) for all of them.
This also fixes the stupid issue with windows computer. Where I had to add a regex to remove `\r`, and that caused a problem because it also removed all blank lines.