Updated the parser
In this update I worked on updating the `parse_blocks` function to remove the need for nested functions and the `this` keyword. This makes the function more generic and removed the need for extra function declarations that lead to confusion because those function names were similar to other functions used in the parser.

Made a major update to how the data is stored.
```
[file.type]: {
 [file.path]: {
  file, // info about the file
  header, // the header comment for the file if it exists
  body // the body comments for the file
 }
}
```